region: add client for AdminService
To send RPCs to HBase that are part of the AdminService, a new
client type is required. For these types of requests, the region
lookup needs to be done beforehand, and the region info provided
as part of the request. This is because the clients can only
handle RPC service, and looking up the region is done using the
ClientService.